+++ Makefile
@@ -51,8 +51,8 @@ vet:
 	go vet ./...
 
 ## staticcheck - Runs static code analyzer staticcheck - currently broken
-# staticcheck: 	
-# 	go run honnef.co/go/tools/cmd/staticcheck@latest ./...
+staticcheck: 	
+	go run honnef.co/go/tools/cmd/staticcheck@latest ./...
 
 ## seccheck - Code vulnerability check
 seccheck:
